feat: merge AI guide into VisitorChat, remove duplicate FAB
- VisitorChat.vue: upgrade FAB from circle to pill-shaped button with label
  * Add glow-pulse breathing animation (fab-glow)
  * Add Tooltip bubble: auto-shows 2s after first visit, localStorage dismiss
  * Add red dot (fab-dot) linked to tooltip visibility
  * Mobile: collapses back to circle icon (<480px)
  * openChat() now also dismisses tooltip

- PublicLayout.vue: remove duplicate AI FAB + modal + all related CSS/script
  * Remove ai-fab-wrap, ai-modal-overlay template blocks
  * Remove aiOpen, showTooltip, openAI, dismissTooltip script logic
  * Remove all .ai-fab-*, .ai-modal-*, animation CSS

Result: single unified AI entry via VisitorChat, no duplicate buttons
